requirements.txt
yfinance
kafka-python
psycopg2-binary
pandas
seaborn
sqlalchemy
apache-airflow=3.1.5